Add initial support for push edges
A push edge between actions A and B means that a push P will be
inserted such that A -vo-> P and P -xo-> B.

There is a bunch of cleanup that ought to be done now that we have
push edges but that I will do later. Now that we have three sorts of
edges, there are a bunch of data structures that ought be arrays
indexed by edge type (instead of just a bunch of seperate struct
fields). Also, we ought to generate push edges from explicit pushes
when using SMT instead of inserting pushes wherever we see them and
then trying to keep track of them.