Add a workaround to allow lambdas to be stored internally with old libstdc++s.
Versions of libstdc++ that predate g++ support for
is_trivially_copyable will only store pointers internally in a
std::function. Workaround this by adding a class that wraps a functor
while manually defining the __is_location_invariant template that
libstdc++ uses to make the decision.

When compiled with a different implementation, everything should still
be fine, but we just might not be accomplishing anything.