Extend the SMT backend with ability to emit release/acquire.
Extend the SMT machinery with release/acquire actions, account for
them in the cost function, extract them, and generate them in the
realizer.  We take advantage of releases in one limited place but
intend to do more.

The goal here is to be able to take advantage of ARMv8's LDA/STR
instructions.