Add a test that exposes a bug in the node freelist based ms queues
Currently they have racy accesses to the elements being sent (since
there is no way to read the data field in a way where it doesn't race
against a write to the node as it is reused). This is obviously
undefined behavior no matter what the underlying type is but when it
is something like std::vector with a highly nontrivial copy
constructor, it can get pretty bad.