Simplify dequeue in the epoch based ms_queues.
Because of the epoch based GC, it is actually no problem if tail falls
behind head. Tail never lags behind in quiescent states: it can only
lag when an enqueuer has not yet updated it. This means that even if
it falls behind head, the node won't be freed until tail has caught up
to head.