Remove unneeded asserts.
Remove an assert in quickEnter() that the Participant is not marked as
exiting in the Participant list. This unfortunately was doing an SC
load in the C11 version, which added a dmb to the quickEnter() path
(which ought to only contain no dmb (or one if we use remote_push
isn't enabled)).