Produce a more tuned C11 epoch implementation.
Previously we only used weak orders on the enter/exit fast path
but the collect path was all SC annotations. This weakens the rest.

The orderings and fences could use some proper documentation but
honestly the way that I produced them is basically that I ran a
variant of the RMC compiler's algorithm in my head on the RMC version
of the code.