Update to make it work on LLVM 12.
This required:
 * Ditching the eliminated BasicBlockPass
 * Changing some function and value names
 * Eliminating uses of the removed TerminatorInst
 * Making sure that all basic blocks have unique names --
   to be honest I'm not sure why this is a /new/ problem
 * Fixing how computeCapacities deals with infinite loops;
   I suspect this was always a problem but that some compilation
   changes maybe made it show up more. (Or maybe it was always busted.)

I didn't bother trying to keep compatability with the old versions
that we supported because it would have been a pain. I'll probably try
to get it building on newer versions too, but I have Ubuntu 20.04
installed on my laptop and it has LLVM 12.

All of the case_studies build and all of the code in examples.  (I
think maybe all the code in examples didn't build in 2017 though, so
that's an improvement.)